+++ diagnostics/finite_amplitude_wave_diag/unused/iMac_model_data_gridfill.py
@@ -4,7 +4,7 @@ Run this on OTC to compute reference state for 1 timestep
 import os
 import numpy as np
 import xarray as xr                # python library we use to read netcdf files
-from diagnostics.finite_amplitude_wave_diag.finite_amplitude_wave_diag_zonal_mean import gridfill_each_level
+from diagnostics.finite_amplitude_wave_diag.finite_amplitude_wave_diag_utils import gridfill_each_level
 from hn2016_falwa.xarrayinterface import QGDataset
 import matplotlib.pyplot as plt
 from hn2016_falwa.oopinterface import QGFieldNHN22
